requirements.txt
# Core dependencies
python-mcp==1.1.2
fastmcp==0.4.0
anthropic==0.42.0
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
pandas==2.2.3

# Development tools
pytest==8.4.0
pytest-asyncio==0.25.2

# Optional but recommended
httpx==0.28.1
websockets==14.1